Stress testing
The idea behind stress testing is that if you have two correct
solutions, and the answer to the problem is unique, then for any
possible test case they are guaranteed to give the same answer. If,
however, at least one of the solutions is incorrect, then with very
high probability there exists a test on which their answers differ.